Final css components before presentation

+++ app/controllers/api/cartitems_controller.rb
@@ -1,36 +1,52 @@
 class Api::CartitemsController < ApplicationController 
+    before_action only: [:index, :update, :destroy]
+
     def index 
-        @cartitems = Cartitem.all
-        render 'api/cartitems/index'
+        if logged_in?
+            @cartitems = Cartitem.all.select{|item| item.user_id == current_user.id }
+            # debugger
+            render 'api/cartitems/index'
+        else
+            return nil
+        end
     end 
 
     def update 
-        @cartitem = Cartitem.find_by(id: params[:id])
-
-        if @cartitem.update 
-            render 'api/cartitems/index'
+        if logged_in?
+            @cartitem = Cartitem.find_by(id: params[:id])
+                if @cartitem.update(cartitem_params)
+                    render 'api/cartitems/index'
+                end
         else
             render json: @cartitem.errors.full_messages, status: 404 
         end
     end 
 
-    def show 
-        @cartitem = Cartitem.find_by(id: params[:id])
-        render 'api/cartitems/show'
-    end
+    # def show 
+    #     @cartitem = Cartitem.all.select(|item| item.user_id == current_user.id ) 
+    #     render 'api/cartitems/show'
+    # end
 
     def destroy 
-        @cartitem = Cartitem.find_by(id: params[:id])
-        if @cartitem.destroy 
-            render'api/cartitems/index'
-        else 
+        if logged_in? 
+            @cartitem = Cartitem.find_by(id: params[:id])
+            # debugger
+            if @cartitem.destroy 
+                # debugger
+                @cartitems = Cartitem.all.select{ |item| item.user_id == current_user.id }
+                # debugger
+                render'api/cartitems/index'
+            end 
+        else
             render json: @cartitem.errors.full_messages, status: 404 
         end
     end
 
     def create 
-        @cartitem = Cartitem.create(cartitem_params)
-        if @cartitem.save 
+        @cartitem = Cartitem.new(cartitem_params)
+        if @cartitem.save && logged_in? 
+            @cartitems = Cartitem.all.select{ |item| item.user_id == current_user.id }
+            # debugger
             render 'api/cartitems/index'
         else 
             render json: @cartitem.errors.full_messages, status: 404
